Create Laundry Expense Tracker.py
This script stores recurring laundry expenses in a text file on Desktop. I enter the paid amount after every laundry session, the script adds it to previously stored value and displays the total cost incurred so far.
It asks if the entry is being for the very first time so that it can create a file with dynamic path that automatically finds out your system username and creates a .txt file on your Desktop folder.